add lntree
and helper script that can be useful in its own right,
make_hard_archive_tree.  The latter's interface is
bad because I wasn't a good enough scripter to easily
make it better at the time I wrote it.

At least I made a test, so that I can fix or rewrite
it in Python* if I decide to work on it more.

Speaking of tests, there's one that needs fixing!

(*probably the language with the most combination
of easy script portability and ability to lessen the filesystem
race conditions; along with Perl, but I know Python better).