add xspf2m3u
The main work was in another person's .xsl file they
shared on the Web; I just made a wrapper shell script
and a test file.  (I don't quickly see a LICENSE file
for their .xsl, but it's so simple, it is the only
way to write an XSLT XSPF-to-M3U converter, that it
might not even qualify for copyright.)

Once upon a time I used this to turn Amarok playlists
into playlists my Sansa Clip could use.  (Later I installed
Rockbox on the Clip and promptly lost the device.)