make keystrokes-tally work with more keyboards
Previously, it was specific to a single model of keyboard.

The new version ends up listening even on weird "keyboards"
like "Power Button" and "Video Bus" (which empirically seems
harmless), because `xinput list` can't easily tell them from
real keyboards.
We try to not include mice (pointer devices) because moving
the mouse generates a lot of events for grep to exclude.
But, some keyboards, like UniversalFoldableKb, show up only
as a "pointer" in `xinput list`, so use a heuristic to find them
(heuristic: lines that contain 'kb' or 'keyboard'
case-insensitively).